map package-json to lock and check in base branch

+++ pkg/scanner/scanner.go
@@ -10,6 +10,7 @@ import (
 	"strings"
 
 	"github.com/pkg/errors"
+	"github.com/samber/lo"
 	"github.com/spf13/viper"
 	"golang.org/x/exp/slices"
 
@@ -74,18 +75,16 @@ func Scan(ctx context.Context, opts flag.Options, cmdName, path string) (*trivyT
 			}
 		}
 
-		var (
-			fileMap map[string]string = nil
-			dir     string
-		)
+		filenameReplaceMap, noLockFiles := oss.GetLockToPackageJson(opts.Target)
 
-		if viper.GetBool("package-json") {
-			dir, fileMap, err = oss.GeneratePackageLockFiles(opts.Target)
+		if viper.GetBool("package-json") && len(noLockFiles) > 0 {
+			dir, newLocksToPackageJson, err := oss.GeneratePackageLockFiles(aquaPath+"/", opts.Target, noLockFiles)
 			if err != nil {
 				log.Logger.Errorf("failed to generate package-lock.json: %s", err)
 			} else {
 				defer os.RemoveAll(dir)
 			}
+			filenameReplaceMap = lo.Assign(filenameReplaceMap, newLocksToPackageJson)
 		}
 
 		// Filesystem scanning
@@ -93,8 +92,8 @@ func Scan(ctx context.Context, opts flag.Options, cmdName, path string) (*trivyT
 			return nil, nil, fmt.Errorf("image scan error: %w", err)
 		}
 
-		if fileMap != nil {
-			fixPackageJsonPaths(&report, fileMap)
+		if filenameReplaceMap != nil {
+			replaceFilenames(&report, filenameReplaceMap)
 		}
 
 	}
@@ -112,7 +111,7 @@ func Scan(ctx context.Context, opts flag.Options, cmdName, path string) (*trivyT
 	return &report, repositoryPipelines, nil
 }
 
-func fixPackageJsonPaths(report *trivyTypes.Report, fileMap map[string]string) {
+func replaceFilenames(report *trivyTypes.Report, fileMap map[string]string) {
 	for i := range report.Results {
 		result := &report.Results[i]
 		if file, ok := fileMap[result.Target]; ok {
